Tools/upgrade_ai/requirements.txt
opencv-python-headless==4.10.0.84
numpy==2.2.1
Pillow==11.0.0
pytz>=2024.2